feat(action): Verify no matching cache before save
Prior to saving cache, verify that there is no cache with a matching key
already cached. Previously, we checked if a cache key matching our cache
was present in the load step; if not, we would then proceed to save the
cache in the save step (assuming we were not in read-only mode and that
there were indeed new Docker images to save). This is problematic if
the action is being run in parallel for multiple jobs; in that case,
there could be multiple, unnecessary cache saves if the cache initially
misses but is subsequently saved.